Edit pagina's klaar
html aanpassing create, edit pagina

+++ HogeschoolPXL/Controllers/HandboekController.cs
@@ -22,20 +22,23 @@ namespace HogeschoolPXL.Controllers
             var handboek = _context.Handboeken;
             return View(handboek);
         }
-        //Nieuwe handboek aanmaken
-        public IActionResult Create()
-        {
-            var handboek = new Handboek();
-            return View(handboek);
-        }
         public IActionResult Details(int id)
         {
             var handboek = _context.Handboeken.Where(x => x.HandboekId == id).FirstOrDefault();
             return View(handboek);
         }
+        //Nieuwe handboek aanmaken
+        [HttpGet]
+        public IActionResult Create()
+        {
+            var handboek = new Handboek();
+            return View(handboek);
+        }
+        //Detail van een Handboek
+        //aanmaken handboek
         [HttpPost]
         public IActionResult Create(Handboek handboek)
-        {
+        {           
             if (ModelState.IsValid)
             {
                 AddHandboek((Handboek)handboek);
@@ -46,12 +49,13 @@ namespace HogeschoolPXL.Controllers
         }
         private void AddHandboek(Handboek handboek)
         {
-            //Gebruiker maken die een lector is
+            //Handboek aanmaken is
             Handboek h = (Handboek)handboek;
             //var h = new Handboek();
             _context.Handboeken.Add(h);
             _context.SaveChanges();
         }
+        //handboek verwijderen
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -59,6 +63,7 @@ namespace HogeschoolPXL.Controllers
                 return NotFound();
             }
             var handboek = await _context.Handboeken.FirstOrDefaultAsync(x => x.HandboekId == id);
+
             if (handboek == null)
             {
                 return NotFound();
@@ -70,7 +75,14 @@ namespace HogeschoolPXL.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var handboek = await _context.Handboeken.FindAsync(id);
+            if (handboek == null)
+            {
+                return NotFound();
+            }
             _context.Handboeken.Remove(handboek);
+            var handboeknaam = handboek.Titel;
+            var handboekid = handboek.HandboekId;
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
